# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API Documentation",
        "description": "REST API for managing Skills",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/v1/profiles": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "List profiles",
                "description": "Returns a paginated list of profiles, including their skills and projects.",
                "operationId": "profileIndex",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of profiles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Profile"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 20
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 97
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Create a profile",
                "description": "Not implemented yet — currently always returns a 501 response.",
                "operationId": "profileStore",
                "responses": {
                    "501": {
                        "description": "Endpoint not implemented",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This endpoint is not implemented yet. The profile creation logic is handled during user registration so far."
                                        },
                                        "code": {
                                            "type": "string",
                                            "example": "NOT_IMPLEMENTED"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/profiles/{profile}": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get a single profile",
                "description": "Returns a profile (route-model-bound by ID) along with its skills and projects.",
                "operationId": "profileShow",
                "parameters": [
                    {
                        "name": "profile",
                        "in": "path",
                        "description": "ID of the profile to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "00000000-0000-0000-0000-000000000000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profile retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Profile"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Profile not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\Models\\Profile]."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Profile"
                ],
                "summary": "Delete a profile",
                "description": "Not implemented yet — currently always returns a 501 response.",
                "operationId": "profileDestroy",
                "parameters": [
                    {
                        "name": "profile",
                        "in": "path",
                        "description": "ID of the profile to delete",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "123e4567-e89b-12d3-a456-426614174000"
                        }
                    }
                ],
                "responses": {
                    "501": {
                        "description": "Endpoint not implemented",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This endpoint is not implemented yet."
                                        },
                                        "code": {
                                            "type": "string",
                                            "example": "NOT_IMPLEMENTED"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update a profile",
                "description": "Validates the request via UpdateProfileRequest and updates the given profile.",
                "operationId": "profileUpdate",
                "parameters": [
                    {
                        "name": "profile",
                        "in": "path",
                        "description": "ID of the profile to update",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "123e4567-e89b-12d3-a456-426614174000"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Fields to update on the profile. TODO: replace with the actual fields from UpdateProfileRequest::rules().",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProfileRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Profile"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Profile not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\Models\\Profile]."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/profiles/{profile}/validate": {
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Validate a profile",
                "description": "Validates the request via ValidateProfileRequest and runs profile validation logic (e.g. moderation/completeness check).",
                "operationId": "profileValidate",
                "parameters": [
                    {
                        "name": "profile",
                        "in": "path",
                        "description": "ID of the profile to validate",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "123e4567-e89b-12d3-a456-426614174000"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Fields required to validate the profile. TODO: replace with the actual fields from ValidateProfileRequest::rules().",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "approved"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile validated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Profile"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Profile not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\Models\\Profile]."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ProfileAccountStatus": {
                "title": "Profile Account Status",
                "description": "The status of the profile",
                "properties": {
                    "PENDING_VALIDATION": {
                        "description": "The profile is pending validation",
                        "type": "string"
                    },
                    "VALIDATED": {
                        "description": "The profile is validated",
                        "type": "string"
                    },
                    "REJECTED": {
                        "description": "The profile is rejected",
                        "type": "string"
                    }
                },
                "type": "string",
                "enum": [
                    "PENDING_VALIDATION",
                    "VALIDATED",
                    "REJECTED"
                ]
            },
...
            "Profile": {
                "title": "Profile",
                "description": "A user profile model",
                "properties": {
                    "id": {
                        "description": "The profile ID",
                        "type": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "user_id": {
                        "description": "The associated user ID",
                        "type": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "email": {
                        "description": "The profile email",
                        "type": "string"
                    },
                    "name": {
                        "description": "The profile name",
                        "type": "string"
                    },
                    "linkedin_id": {
                        "description": "The profile LinkedIn ID",
                        "type": "string"
                    },
                    "headline": {
                        "description": "The profile headline",
                        "type": "string"
                    },
                    "bio": {
                        "description": "The profile bio",
                        "type": "string"
                    },
                    "avatar_url": {
                        "description": "The profile avatar URL",
                        "type": "string"
                    },
                    "location": {
                        "description": "The profile location",
                        "type": "string"
                    },
                    "status": {
                        "description": "The profile status",
                        "type": "string"
                    },
                    "account_status": {
                        "$ref": "#/components/schemas/ProfileAccountStatus"
                    },
                    "skills": {
                        "description": "Skills linked to this profile",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Skill"
                        }
                    },
                    "projects": {
                        "description": "Projects linked to this profile",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Project"
                        }
                    },
                    "created_at": {
                        "description": "The creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "The update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Project": {
                "title": "Project",
                "description": "A project model associated with a user profile",
                "properties": {
                    "id": {
                        "description": "The project ID",
                        "type": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "title": {
                        "description": "The project title",
                        "type": "string"
                    },
                    "description": {
                        "description": "The project description",
                        "type": "string"
                    },
                    "link": {
                        "description": "The project link",
                        "type": "string"
                    },
                    "profile_id": {
                        "description": "The associated profile ID",
                        "type": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "created_at": {
                        "description": "The creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "The update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Skill": {
                "title": "Skill",
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "name": {
                        "type": "string",
                        "example": "Laravel"
                    },
                    "slug": {
                        "type": "string",
                        "example": "laravel"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}